requirements.txt
asgiref==3.5.2
Django==4.0.6
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
